pyproject.toml
# File: pyproject.toml
[project]
name = "biomesh"
version = "1.0.0"
description = "Biotech multi-agent research platform implementing Google A2A protocol"
requires-python = ">=3.11,<3.14"
dependencies = [
    "langchain==0.3.7",
    "langchain-community==0.3.7",
    "langchain-groq==0.2.1",
    "fastapi==0.115.5",
    "uvicorn==0.32.1",
    "python-dotenv==1.0.1",
    "httpx==0.27.2",
    "gradio==4.44.1",
    "sse-starlette==2.1.3",
    "pydantic==2.9.2",
    "arxiv==2.1.0",
    "requests==2.31.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["a2a", "agents", "app"]

[tool.uv]
